requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
openai
Flask-Limiter
Flask-Caching
firebase-admin
requests
weasyprint
openpyxl
bleach
tinycss2
Pillow>=10.0.0